Move ProcessSteps into separate package
