+++ compiler/src/main/kotlin/plumb/internal/codegen/PlumbProcessor.kt
@@ -4,6 +4,7 @@ import com.google.auto.service.AutoService
 import plumb.annotation.In
 import plumb.annotation.Out
 import plumb.annotation.Plumbed
+import plumb.internal.codegen.step.ProcessSteps
 import javax.annotation.processing.AbstractProcessor
 import javax.annotation.processing.Filer
 import javax.annotation.processing.Messager
