--- compiler/src/main/kotlin/plumb/internal/codegen/ProcessStep.kt
@@ -1,5 +1,0 @@
-package plumb.internal.codegen
-
-interface ProcessStep {
-    fun process(model: Model)
-}
