+++ compiler/src/main/kotlin/plumb/internal/codegen/step/ProcessStep.kt
@@ -1,0 +1,7 @@
+package plumb.internal.codegen.step
+
+import plumb.internal.codegen.Model
+
+interface ProcessStep {
+    fun process(model: Model)
+}
